refactor: extract FlowStepCard from FlowEditor
- new runes-native src/components/FlowStepCard.svelte owns the step
  card row, run-status badges, and the override panel (alias, URL,
  headers, Body/Assertions/Before Send/After Receive tabs); card CSS
  moves with it
- FlowEditor keeps flow-level state (uiState, alias drafts, variable
  picker, save plumbing) and passes per-step callbacks down; the shared
  DragReorder controller is passed to each card
- directivesToText moves to src/lib/pbScript.ts so both components can
  share it
- FlowEditor 2028 -> 882 lines
